requirements.txt
langchain
weaviate-client
langchain_community
huggingface_hub
langchain_huggingface
transformers
streamlit
einops
re